move code to plugin directory
