+++ test/plugins/signoff/SimpleReview/SimpleReview_test.tsx
@@ -2,13 +2,13 @@ import { expect } from "chai";
 
 import testUtils from "react-dom/test-utils";
 
-import { createComponent } from "../../test_utils";
+import { createComponent } from "../../../test_utils";
 import SimpleReview, {
   SimpleReviewProps,
-} from "../../../src/components/simpleReview/SimpleReview";
+} from "../../../../src/plugins/signoff/SimpleReview/SimpleReview";
 
-import { SessionState } from "../../../src/types";
-import { SignoffState } from "../../../src/plugins/signoff/types";
+import { SessionState } from "../../../../src/types";
+import { SignoffState } from "../../../../src/plugins/signoff/types";
 
 function signoffFactory(): SignoffState {
   return {
